Add files via upload
AW Scheduler is multi-user, multi-customer appointment scheduler I developed as a university project for Advanced C# - c969. It demonstrates the utilization of a MySQL database in management of data related to users, customers, and appointments relative to the current user's local time zone. The login screen displays in English or Spanish depending on the user's computer's selected region.